Create README.md
modNFTs.com, KYIAM.Univer.se and NFTeedIT.com are 
founding members of our fast growing community 
created for helping one another as much as possible.
We are always open to receiving knowledge from
our peers who are speaking from experience. We 
appreciate your time and are Thankful to all for 
helping us to educate ourselves and people about 
the sometimes overwhelming amounts of information 
and oppurtunities available throughout this vast 
ever expanding space we call The Internet. 

Everyone is Welcome to Join
or if you are simply checking things out and 
a bit unsure about joining we understand. 
Joining or not, all can feel free to ask anything 
and you should get a response within a day or two.
Have a great day. Sincerely, Mod1 & 3Sons.